requirements.txt
python-telegram-bot==20.3
yt-dlp==2024.5.27
requests==2.31.0
